third_party/WebKit/LayoutTests/screen_orientation/screenorientation-api-expected.txt

Validates the Screen Orientation API is exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.lockOrientation is defined.
PASS screen.unlockOrientation is defined.
PASS screen.orientation is defined.
FAIL window.onorientationchange should be defined. Was undefined
PASS successfullyParsed is true

TEST COMPLETE
